Refactor GeoPoint and GeoPointBinding packages: move from boundary to entity and update references
